# apps/cli/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Quotio Local HTTP API",
    "version": "1"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:6767"
    }
  ],
  "paths": {
...
    "/v2/discovery": {
      "get": {
        "summary": "Read host native discovery state",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Native discovery state",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2Discovery"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Scan and register native sources on the host",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                  "restore_removed": {"type":"boolean", "default":false, "description":"Explicitly reconnect previously removed native sources for the selected providers."},
                  "providers": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/ProviderId"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Discovery operation; poll its operation ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Operation"
                }
              }
            }
          }
        }
      }
    },
    "/v2/snapshot": {
      "get": {
        "summary": "Read resolved account, source health and usage state",
        "security": [{"bearerAuth": []}, {}],
        "responses": {
          "200": {"description": "Consistent resolved snapshot", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/V2Snapshot"}}}},
          "503": {"description": "Host account state is unavailable"}
        }
      }
    },
...
    "/v2/accounts": {
      "get": {
        "summary": "Read resolved accounts without provider network requests",
        "description": "Reads the shared resolved account model. Protected metadata is initialized atomically on first access; repeated reads do not change IDs or revision. No native provider credential scan or provider HTTP request is performed.",
        "security": [{"bearerAuth": []}, {}],
        "responses": {
          "200": {
            "description": "Resolved account list shared with the default CLI account list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2AccountList"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          }
        }
      },
      "post": {
        "summary": "Create a monitored account using the existing typed credential intake",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Idempotent operation; poll /v2/operations/{id}",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Operation"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/ApiKeyInput"
                  },
                  {
                    "$ref": "#/components/schemas/GrokOwnedInput"
                  },
                  {
                    "$ref": "#/components/schemas/FactoryOwnedInput"
                  },
                  {
                    "$ref": "#/components/schemas/KiroOwnedInput"
                  },
                  {
                    "$ref": "#/components/schemas/AntigravityOwnedInput"
                  }
                ]
              }
            }
          }
        }
      }
    },
    "/openapi.json": {
      "get": {
        "summary": "OpenAPI document",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "405": {
            "$ref": "#/components/responses/MethodNotAllowed"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          },
          "500": {
            "$ref": "#/components/responses/BadRequest",
            "description": "Internal error"
          }
        },
        "security": [
          {},
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/v2/status": {
      "get": {
        "summary": "Refresh and settings status",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Status"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "405": {
            "$ref": "#/components/responses/MethodNotAllowed"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          },
          "500": {
            "$ref": "#/components/responses/BadRequest",
            "description": "Internal error"
          }
        },
        "security": [
          {},
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/v2/providers": {
      "get": {
        "summary": "List provider capabilities",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderList"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "405": {
            "$ref": "#/components/responses/MethodNotAllowed"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          },
          "500": {
            "$ref": "#/components/responses/BadRequest",
            "description": "Internal error"
          }
        },
        "security": [
          {},
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/v2/settings": {
      "get": {
        "summary": "Get settings",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsView"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "405": {
            "$ref": "#/components/responses/MethodNotAllowed"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          },
          "500": {
            "$ref": "#/components/responses/BadRequest",
            "description": "Internal error"
          }
        },
        "security": [
          {},
          {
            "bearerAuth": []
          }
        ]
      },
      "patch": {
        "summary": "Patch settings",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsView"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "405": {
            "$ref": "#/components/responses/MethodNotAllowed"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          },
          "500": {
            "$ref": "#/components/responses/BadRequest",
            "description": "Internal error"
          },
          "408": {
            "$ref": "#/components/responses/BadRequest",
            "description": "Request body timeout"
          },
          "413": {
            "$ref": "#/components/responses/BadRequest",
            "description": "JSON body exceeds 64 KiB"
          },
          "415": {
            "$ref": "#/components/responses/BadRequest",
            "description": "JSON content type required"
          },
          "422": {
            "$ref": "#/components/responses/BadRequest",
            "description": "Invalid JSON request"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettingsPatch"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Invalid request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Unauthorized",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "NotFound": {
        "description": "Not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "Conflict": {
        "description": "Conflict",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Unavailable": {
        "description": "Unavailable",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    },
    "schemas": {
...
      "V2Discovery": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "schema_version",
          "scans",
          "permissions",
          "known_sources",
          "failures",
          "registered"
        ],
        "properties": {
          "schema_version": {
            "const": 2
          },
          "scans": {
            "type": "array",
            "items": {
              "type": "object",
              "additionalProperties": false,
              "required": [
                "provider",
                "at"
              ],
              "properties": {
                "provider": {
                  "$ref": "#/components/schemas/ProviderId"
                },
                "at": {
                  "type": "string",
                  "format": "date-time"
                }
              }
            }
          },
          "permissions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/V2NativePermission"
            }
          },
          "known_sources": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/V2NativePermission"
            }
          },
          "failures": {
            "type": "array",
            "items": {
              "type": "object",
              "additionalProperties": false,
              "required": [
                "provider",
                "kind",
                "code"
              ],
              "properties": {
                "provider": {
                  "$ref": "#/components/schemas/ProviderId"
                },
                "kind": {
                  "type": "string"
                },
                "code": {
                  "type": "string"
                }
              }
            }
          },
          "registered": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
...
      "V2AccountList": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "schema_version",
          "host",
          "revision",
          "accounts"
        ],
        "properties": {
          "account_redirects": {"type": "object", "additionalProperties": {"type": "string"}},
          "schema_version": {
            "const": 2
          },
          "host": {
            "$ref": "#/components/schemas/V2Host"
          },
          "revision": {
            "type": "integer",
            "minimum": 0
          },
          "accounts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/V2Account"
            }
          }
        }
      },
...
      "ProviderId": {
        "type": "string",
        "enum": [
          "aiand",
          "alibabacodingplan",
          "amp",
          "antigravity",
          "azureopenai",
          "chutes",
          "claude",
          "clawrouter",
          "clinepass",
          "codebuff",
          "codex",
          "copilot",
          "crof",
          "cursor",
          "deepgram",
          "deepinfra",
          "deepseek",
          "devin",
          "devin-desktop",
          "doubao",
          "elevenlabs",
          "factory",
          "fireworks",
          "gemini",
          "grok",
          "groq",
          "ibmbob",
          "kilo",
          "kimi",
          "kiro",
          "litellm",
          "llmproxy",
          "minimax",
          "mock",
          "moonshot",
          "neuralwatt",
          "openai",
          "opencodego",
          "openrouter",
          "poe",
          "sub2api",
          "synthetic",
          "venice",
          "vertexai",
          "warp",
          "xai",
          "zai",
          "zenmux"
        ]
      },
...
      "Operation": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "kind": {
            "type": "string",
            "enum": [
              "refresh",
              "account_create",
              "account_update",
              "account_remove",
              "usage_query"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "running",
              "completed",
              "failed"
            ]
          },
          "result": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "type": "object",
                "additionalProperties": false,
                "required": [
                  "account_id"
                ],
                "properties": {
                  "account_id": {
                    "type": "string"
                  }
                }
              },
              {
                "type": "object",
                "additionalProperties": false,
                "required": [
                  "providers",
                  "failures"
                ],
                "properties": {
                  "providers": {
                    "type": "integer",
                    "minimum": 0
                  },
                  "failures": {
                    "type": "integer",
                    "minimum": 0
                  }
                }
              }
            ]
          },
          "error": {
            "type": [
              "string",
              "null"
            ]
          },
          "message": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "additionalProperties": false,
        "required": [
          "id",
          "kind",
          "status",
          "result",
          "error",
          "message"
        ],
        "description": "At most 128 operations run concurrently. The latest 128 completed refresh results expire after 15 minutes. Account write results and idempotency keys persist until restart, limited to 4096 distinct keys; new keys beyond that limit return 503 idempotency_full. Existing keys remain replayable."
      },
      "ApiKeyInput": {
        "type": "object",
        "properties": {
          "provider": {
            "$ref": "#/components/schemas/ProviderId"
          },
          "label": {
            "type": [
              "string",
              "null"
            ]
          },
          "api_key": {
            "type": "string",
            "minLength": 1,
            "maxLength": 16384
          },
          "settings": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "region": {
            "type": [
              "string",
              "null"
            ]
          },
          "organization": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "additionalProperties": false,
        "required": [
          "provider",
          "api_key",
          "settings",
          "region",
          "organization"
        ]
      },
...
      "AntigravityOwnedInput": {
        "type": "object",
        "additionalProperties": false,
        "required": ["kind", "label", "access_token", "refresh_token", "expires_at", "client_id", "client_secret"],
        "properties": {
          "kind": {"const": "antigravity_owned"},
          "label": {"type": "string", "minLength": 1, "maxLength": 100},
          "access_token": {"type": "string", "writeOnly": true},
          "refresh_token": {"type": "string", "writeOnly": true},
          "expires_at": {"type": "integer"},
          "client_id": {"const": "1071006060591-tmhssin2h21lcre235vtolojh4g403ep.apps.googleusercontent.com"},
          "client_secret": {"type": "string", "writeOnly": true}
        },
        "description": "Separately owned Google OAuth tokens for the supported Antigravity client. The caller must supply the client secret explicitly; Quotio does not ship a secret or extract one from an installed app. Inputs stay inside the protected vault. Refresh uses only https://oauth2.googleapis.com/token. No caller endpoint, browser login, native import, or proxy discovery. Refresh is durably marked pending before sending; rotation is saved before quota. An interrupted or uncertain refresh cannot be replayed. Refresh-token reservations survive rotation and deletion. Vault format 8 rejects older readers and downgraded documents containing these credentials or reservations."
      },
...
      "GrokOwnedInput": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "kind",
          "label",
          "access_token",
          "refresh_token",
          "expires_at"
        ],
        "properties": {
          "kind": {
            "const": "grok_owned"
          },
          "label": {
            "type": "string",
            "minLength": 1,
            "maxLength": 80
          },
          "access_token": {
            "type": "string",
            "minLength": 1,
            "maxLength": 16384,
            "writeOnly": true
          },
          "refresh_token": {
            "type": "string",
            "minLength": 1,
            "maxLength": 16384,
            "writeOnly": true
          },
          "expires_at": {
            "type": "integer",
            "minimum": 0
          }
        },
        "description": "Explicit intake of separately owned Grok CLI OAuth tokens for the default Grok client. Does not read or migrate native credentials or start OAuth. Tokens are stored before any refresh. An interrupted or failed refresh is fenced until this account is removed and separately owned credentials are supplied again."
      },
...
      "V2Snapshot": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "schema_version",
          "host",
          "revision",
          "generated_at",
          "accounts",
          "usage"
        ],
        "properties": {
          "account_redirects": {"type": "object", "additionalProperties": {"type": "string"}},
          "provider_issues": {"type": "object", "additionalProperties": {"$ref": "#/components/schemas/V2Issue"}},
          "schema_version": {
            "const": 2
          },
          "host": {
            "$ref": "#/components/schemas/V2Host"
          },
          "revision": {
            "type": "integer",
            "minimum": 0
          },
          "generated_at": {
            "type": "string",
            "format": "date-time"
          },
          "accounts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/V2Account"
            }
          },
          "usage": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/V2Usage"
            }
          }
        },
        "description": "Draft full snapshot contract. The opt-in /v2/accounts read API is available; a full snapshot endpoint and v2 CRUD are not yet exposed."
      }
    }
  },
...
}